+++ crates/api-core/src/tests/common/api_fixtures/mod.rs
@@ -1663,7 +1663,9 @@ pub async fn create_test_env_with_overrides(
                 per_object_metrics_registry: per_object_metrics_registry.clone(),
                 rack_firmware_reprovisioning_enabled: false,
                 redfish_client_pool: redfish_sim.clone(),
-                bmc_rotation_gate: carbide_credential_rotation::BmcRotationGate::new(),
+                bmc_rotation_gate: carbide_credential_rotation::RotationGate::new_for_family(
+                    db::credential_rotation::CredentialRotationType::Bmc,
+                ),
                 bmc_rotation_enabled: false,
             }
             .into(),
