+++ crates/api-core/src/tests/power_shelf_state_controller/reprovisioning.rs
@@ -56,7 +56,9 @@ fn services(
         per_object_metrics_registry: env.per_object_metrics_registry(),
         rack_firmware_reprovisioning_enabled,
         redfish_client_pool: env.redfish_sim.clone(),
-        bmc_rotation_gate: carbide_credential_rotation::BmcRotationGate::new(),
+        bmc_rotation_gate: carbide_credential_rotation::RotationGate::new_for_family(
+            db::credential_rotation::CredentialRotationType::Bmc,
+        ),
         bmc_rotation_enabled: false,
     }
 }
