+++ crates/api-model/src/machine/mod.rs
@@ -840,11 +840,11 @@ pub struct Machine {
     /// bypassing the passive site-wide gate and the device's backoff quarantine.
     pub bmc_credential_rotation_requested: bool,
 
-    /// Operator "force-converge this UEFI credential now" request (REQ-2, UEFI).
+    /// Operator "force-converge this UEFI credential now" request.
     /// Set on the machine that owns the UEFI credential (a host machine for its
-    /// host UEFI; a DPU machine for its DPU UEFI in the follow-up PR). When
-    /// `true`, the machine state controller enters `RotatingHostUefi` (or, in the
-    /// follow-up PR, `RotatingDpuUefi`) and force-converges this machine's UEFI
+    /// host UEFI; a DPU machine for its DPU UEFI). When
+    /// `true`, the machine state controller enters `RotatingHostUefi` (or
+    /// `RotatingDpuUefi` for a DPU) and force-converges this machine's UEFI
     /// credential on its next sweep,
     /// bypassing the passive site-wide gate and the device's backoff quarantine.
     pub uefi_credential_rotation_requested: bool,
@@ -950,6 +950,15 @@ impl Machine {
             Some(hw) => hw.bmc_vendor(),
             None => bmc_vendor::BMCVendor::Unknown,
         }
+    }
+
+    /// Whether this machine's BMC must have its lockdown disabled before its
+    /// UEFI (BIOS setup) password can be changed, and re-enabled afterward.
+    /// Only Dell BMCs enforce a lockdown that blocks the BIOS password change;
+    /// other vendors accept it without unlocking. Centralizes the vendor check
+    /// so the UEFI setup and rotation flows don't special-case Dell inline.
+    pub fn needs_bmc_unlock_for_uefi_setup(&self) -> bool {
+        self.bmc_vendor().is_dell()
     }
 
     /// Does the forge-dpu-agent on this DPU need upgrading?
@@ -1290,7 +1299,7 @@ pub enum ManagedHostState {
     },
 
     /// The host is converging its own UEFI (BIOS setup) password to the staged
-    /// site-wide rotation target (REQ-2, UEFI). A pool-only, top-level state: it
+    /// site-wide rotation target. A pool-only, top-level state: it
     /// blocks instance creation (which requires exact `Ready`) for the bounded
     /// duration of the rotation. Unlike `RotatingBmc`, applying a new UEFI
     /// password requires a BIOS config job plus a full host power-cycle, so this
@@ -1303,8 +1312,8 @@ pub enum ManagedHostState {
     /// This state is host-specific on purpose: a DPU's UEFI password is a
     /// distinct device (keyed by the DPU BMC MAC), applied through a DPU restart
     /// rather than a host power-cycle, and a host can carry several DPUs. That
-    /// gets its own sibling `RotatingDpuUefi` state in the follow-up PR rather
-    /// than an overloaded discriminator here.
+    /// gets its own sibling `RotatingDpuUefi` state rather than an overloaded
+    /// discriminator here.
     RotatingHostUefi {
         uefi_setup_info: UefiSetupInfo,
     },
