+++ crates/api-model/src/machine/slas.rs
@@ -68,7 +68,7 @@ pub const MAINTENANCE: Duration = Duration::from_secs(5 * 60);
 // engine's short per-device backoff without tripping the SLA on the first retry.
 pub const ROTATING_BMC: Duration = Duration::from_secs(15 * 60);
 
-// Host UEFI credential rotation (REQ-2, UEFI). Applying a new UEFI password
+// Host UEFI credential rotation. Applying a new UEFI password
 // requires a BIOS config job plus a full host power-cycle and job-completion
 // poll, so the budget is more generous than the synchronous BMC rotation: enough
 // to absorb a slow reboot plus the engine's short per-device backoff without
