+++ crates/power-shelf-controller/src/rotating_bmc.rs
@@ -25,7 +25,7 @@
 //!
 //! - *Should we enter rotation?* [`should_enter_bmc_rotation`] honors an operator
 //!   force-converge request first (even when the site-wide flag is off), then the
-//!   passive gate (`bmc_rotation_enabled` && the cached [`BmcRotationGate`] says
+//!   passive gate (`bmc_rotation_enabled` && the cached [`RotationGate`] says
 //!   the PMC lags the staged target).
 //! - *Do one rotation tick.* [`handle_rotating_bmc`] converges the PMC via
 //!   [`rotate_bmc`], folds the outcome through the shared state-neutral
@@ -75,7 +75,7 @@ pub async fn should_enter_bmc_rotation(
     };
     services
         .bmc_rotation_gate
-        .bmc_rotation_needed(&services.db_pool, endpoint.device_mac)
+        .rotation_needed(&services.db_pool, endpoint.device_mac)
         .await
         .map_err(|e| StateHandlerError::GenericError(eyre::eyre!("bmc rotation gate query: {e}")))
 }
